+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - added option `exitCallbackOptions` to configure an array of callback options executed on click a new secondary button
   - added option `secondaryCallbackOptions` to configure an array of callback options executed on click of the primary (exit) button
   - added option `secondaryButtonTitle` to configure the new secondary button title
+- Reimplements VcsTreeview without the VTreeview component; significantly improved performance.
 
 # 6.0.11